+++ vertx-mysql-client/src/test/java/io/vertx/mysqlclient/MySQLQueryTest.java
@@ -48,14 +48,14 @@ public class MySQLQueryTest extends MySQLTestBase {
   public void testLastInsertIdWithDefaultValue(TestContext ctx) {
     MySQLConnection.connect(vertx, options, ctx.asyncAssertSuccess(conn -> {
       conn.query("CREATE TEMPORARY TABLE last_insert_id(id INTEGER PRIMARY KEY AUTO_INCREMENT, val VARCHAR(20));", ctx.asyncAssertSuccess(createTableResult -> {
-        int lastInsertId1 = createTableResult.property(MySQLClient.LAST_INSERTED_ID);
-        ctx.assertEquals(0, lastInsertId1);
+        long lastInsertId1 = createTableResult.property(MySQLClient.LAST_INSERTED_ID);
+        ctx.assertEquals(0L, lastInsertId1);
         conn.query("INSERT INTO last_insert_id(val) VALUES('test')", ctx.asyncAssertSuccess(insertResult1 -> {
-          int lastInsertId2 = insertResult1.property(MySQLClient.LAST_INSERTED_ID);
-          ctx.assertEquals(1, lastInsertId2);
+          long lastInsertId2 = insertResult1.property(MySQLClient.LAST_INSERTED_ID);
+          ctx.assertEquals(1L, lastInsertId2);
           conn.query("INSERT INTO last_insert_id(val) VALUES('test2')", ctx.asyncAssertSuccess(insertResult2 -> {
-            int lastInsertId3 = insertResult2.property(MySQLClient.LAST_INSERTED_ID);
-            ctx.assertEquals(2, lastInsertId3);
+            long lastInsertId3 = insertResult2.property(MySQLClient.LAST_INSERTED_ID);
+            ctx.assertEquals(2L, lastInsertId3);
             conn.close();
           }));
         }));
@@ -67,17 +67,17 @@ public class MySQLQueryTest extends MySQLTestBase {
   public void testLastInsertIdWithSpecifiedValue(TestContext ctx) {
     MySQLConnection.connect(vertx, options, ctx.asyncAssertSuccess(conn -> {
       conn.query("CREATE TEMPORARY TABLE last_insert_id(id INTEGER PRIMARY KEY AUTO_INCREMENT, val VARCHAR(20));", ctx.asyncAssertSuccess(createTableResult -> {
-        int lastInsertId1 = createTableResult.property(MySQLClient.LAST_INSERTED_ID);
-        ctx.assertEquals(0, lastInsertId1);
+        long lastInsertId1 = createTableResult.property(MySQLClient.LAST_INSERTED_ID);
+        ctx.assertEquals(0L, lastInsertId1);
         conn.query("ALTER TABLE last_insert_id AUTO_INCREMENT=1234", ctx.asyncAssertSuccess(alterTableResult -> {
-          int lastInsertId2 = createTableResult.property(MySQLClient.LAST_INSERTED_ID);
-          ctx.assertEquals(0, lastInsertId2);
+          long lastInsertId2 = createTableResult.property(MySQLClient.LAST_INSERTED_ID);
+          ctx.assertEquals(0L, lastInsertId2);
           conn.query("INSERT INTO last_insert_id(val) VALUES('test')", ctx.asyncAssertSuccess(insertResult1 -> {
-            int lastInsertId3 = insertResult1.property(MySQLClient.LAST_INSERTED_ID);
-            ctx.assertEquals(1234, lastInsertId3);
+            long lastInsertId3 = insertResult1.property(MySQLClient.LAST_INSERTED_ID);
+            ctx.assertEquals(1234L, lastInsertId3);
             conn.query("INSERT INTO last_insert_id(val) VALUES('test2')", ctx.asyncAssertSuccess(insertResult2 -> {
-              int lastInsertId4 = insertResult2.property(MySQLClient.LAST_INSERTED_ID);
-              ctx.assertEquals(1235, lastInsertId4);
+              long lastInsertId4 = insertResult2.property(MySQLClient.LAST_INSERTED_ID);
+              ctx.assertEquals(1235L, lastInsertId4);
               conn.close();
             }));
           }));
